+++ tests/testthat/test_plot_inspect_type.R
@@ -14,19 +14,16 @@ test_that("inspect_types plots", {
   expect_doppelganger("Inspect-types-tech",   tech %>%   inspect_types %>% show_plot)
 })
 
-# test_that("inspect_types plot paired", {
-#   expect_doppelganger("Inspect-types-paired-starwars", starwars %>% inspect_types(x2) %>% show_plot)
-#   expect_doppelganger("Inspect-types-paired-storms",   storms %>%   inspect_types(y2) %>% show_plot)
-# })
-
 test_that("inspect_types plot suppress labels", {
   expect_doppelganger("Inspect-types-no-lab-starwars", starwars %>% inspect_types %>% show_plot(text_labels = FALSE))
   expect_doppelganger("Inspect-types-no-lab-tech", tech %>% inspect_types %>% show_plot(text_labels = FALSE))
   expect_doppelganger("Inspect-types-no-lab-storms",   storms %>%   inspect_types %>% show_plot(text_labels = FALSE))
 })
 
-test_that("inspect_types plot color palette", {
-  expect_doppelganger("Inspect-types-palette-starwars", starwars %>% inspect_types() %>% show_plot(col_palette = 1))
-  expect_doppelganger("Inspect-types-palette-storms",   storms %>%   inspect_types() %>% show_plot(col_palette = 2))
+test_that("inspect_types plot paired", {
+  expect_doppelganger("Inspect-types-paired-starwars", starwars %>% inspect_types(x2) %>% show_plot)
+  expect_doppelganger("Inspect-types-paired-storms",   storms %>%   inspect_types(y2) %>% show_plot)
 })
+
+